effect-bdd: organize src into core/ subfolder
Move the five generic DSL modules and their tests into src/core/ (mirroring the
cypher-codegen subfolder convention), leaving the barrel at src/index.ts pointing
to ./core/*. The former Bdd test becomes core/Interpreting (one scenario, many
interpreters). Public API and path alias unchanged.
