+++ packages/bdd/src/core/Interpreting.node.unit.test.ts
@@ -1,8 +1,5 @@
 import { describe, expect, it } from "@effect/vitest"
 import { Effect } from "effect"
-import * as fs from "node:fs"
-import * as path from "node:path"
-import { fileURLToPath } from "node:url"
 import { steps, tags, toGherkin } from "./Gherkin.js"
 import { run } from "./Run.js"
 import { scenario } from "./Scenario.js"
@@ -26,19 +23,5 @@ describe("interpreting one scenario many ways", () => {
   it("exposes the same value as structure", () => {
     expect(steps(cart).map((step) => step.keyword)).toEqual(["Given", "When", "Then"])
     expect(tags(cart)).toEqual(["cart"])
-  })
-})
-
-describe("runner-agnostic core", () => {
-  it("imports no test runner from non-test source", () => {
-    const sourceDir = fileURLToPath(new URL(".", import.meta.url))
-    const sources = fs
-      .readdirSync(sourceDir)
-      .filter((file) => file.endsWith(".ts") && !file.endsWith(".test.ts") && !file.endsWith(".test-d.ts"))
-
-    const runnerImport = /from\s+["'](?:vitest|@effect\/vitest|jest|@jest\/globals|mocha|node:test|bun:test)["']/
-
-    const offenders = sources.filter((file) => runnerImport.test(fs.readFileSync(path.join(sourceDir, file), "utf8")))
-    expect(offenders).toEqual([])
   })
 })
